moved test code.
